formateo con mvn formatter:validate y reorganización de archivos en repository
