+++ src/main/java/tacs/eventos/EventosApplication.java
@@ -1,6 +1,5 @@
 package tacs.eventos;
 
-import org.modelmapper.ModelMapper;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,7 +8,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import tacs.eventos.model.RolUsuario;
 import tacs.eventos.model.Usuario;
-import tacs.eventos.repository.UsuarioInMemoryRepository;
+import tacs.eventos.repository.usuario.UsuarioInMemoryRepository;
 
 import java.util.Set;
