+++ src/main/java/tacs/eventos/config/ModelMapperConfig.java
@@ -12,20 +12,11 @@ public class ModelMapperConfig {
     @Bean
     public ModelMapper modelMapper() {
         ModelMapper mapper = new ModelMapper();
-        mapper.createTypeMap(EventoDTO.class, Evento.class)
-                .setProvider(request -> {
-                    EventoDTO dto = (EventoDTO) request.getSource(); // casteo a EventoDTO
-                    return new Evento(
-                            dto.titulo(),
-                            dto.descripcion(),
-                            dto.fechaHoraInicio(),
-                            dto.duracionMinutos(),
-                            dto.ubicacion(),
-                            dto.cupoMaximo(),
-                            dto.precio(),
-                            dto.categoria()
-                    );
-                });
+        mapper.createTypeMap(EventoDTO.class, Evento.class).setProvider(request -> {
+            EventoDTO dto = (EventoDTO) request.getSource(); // casteo a EventoDTO
+            return new Evento(dto.titulo(), dto.descripcion(), dto.fechaHoraInicio(), dto.duracionMinutos(),
+                    dto.ubicacion(), dto.cupoMaximo(), dto.precio(), dto.categoria());
+        });
         return mapper;
     }
 }
