+++ src/main/java/tacs/eventos/controller/EventoController.java
@@ -21,9 +21,10 @@ public class EventoController {
 
     @PostMapping
     public Evento crearEvento(@RequestBody EventoDTO dto) {
-            Evento evento = mapper.map(dto, Evento.class);
-            return eventoService.crearEvento(evento);
+        Evento evento = mapper.map(dto, Evento.class);
+        return eventoService.crearEvento(evento);
     }
+
     @GetMapping
     public List<Evento> listarEventos() {
         return eventoService.listarEventos();
