+++ src/main/java/tacs/eventos/repository/evento/EventosInMemoryRepo.java
@@ -1,4 +1,4 @@
-package tacs.eventos.repository;
+package tacs.eventos.repository.evento;
 
 import org.springframework.stereotype.Repository;
 import tacs.eventos.model.Evento;
