+++ src/main/java/tacs/eventos/repository/sesion/SessionInMemoryRepository.java
@@ -1,4 +1,4 @@
-package tacs.eventos.repository;
+package tacs.eventos.repository.sesion;
 
 import org.springframework.stereotype.Repository;
 import tacs.eventos.model.Session;
