+++ src/main/java/tacs/eventos/service/SessionService.java
@@ -5,8 +5,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import tacs.eventos.model.Session;
 import tacs.eventos.model.Usuario;
-import tacs.eventos.repository.SessionRepository;
-import tacs.eventos.repository.UsuarioRepository;
+import tacs.eventos.repository.sesion.SessionRepository;
+import tacs.eventos.repository.usuario.UsuarioRepository;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
