+++ src/test/java/tacs/eventos/repository/EventosInMemoryRepoTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import tacs.eventos.model.Evento;
+import tacs.eventos.repository.evento.EventosInMemoryRepo;
 
 @SpringBootTest
 class EventosInMemoryRepoTest {
