+++ src/test/java/tacs/eventos/service/UsuarioServiceTest.java
@@ -2,15 +2,14 @@ package tacs.eventos.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import tacs.eventos.dto.EventoDTO;
 import tacs.eventos.model.Evento;
 import tacs.eventos.model.InscripcionEvento;
 import tacs.eventos.model.RolUsuario;
 import tacs.eventos.model.Usuario;
-import tacs.eventos.repository.InscripcionesRepository;
-import tacs.eventos.repository.UsuarioRepository;
+import tacs.eventos.repository.inscripcion.InscripcionesRepository;
+import tacs.eventos.repository.usuario.UsuarioRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -57,7 +56,8 @@ class UsuarioServiceTest {
         String email = "test@example.com";
         String password = "password123";
 
-        when(usuarioRepository.obtenerPorEmail(email)).thenReturn(Optional.of(new Usuario(email, password, Set.of(RolUsuario.USUARIO))));
+        when(usuarioRepository.obtenerPorEmail(email))
+                .thenReturn(Optional.of(new Usuario(email, password, Set.of(RolUsuario.USUARIO))));
 
         Exception exception = assertThrows(IllegalArgumentException.class, () -> {
             usuarioService.registrar(email, password);
@@ -85,10 +85,8 @@ class UsuarioServiceTest {
         Evento evento1 = new Evento("Evento 1", "Desc 1", null, 60, "Ubicacion", 100, 500, "Categoria");
         Evento evento2 = new Evento("Evento 2", "Desc 2", null, 120, "Ubicacion", 50, 1000, "Categoria");
 
-        List<InscripcionEvento> inscripciones = List.of(
-                new InscripcionEvento(usuarioId, evento1),
-                new InscripcionEvento(usuarioId, evento2)
-        );
+        List<InscripcionEvento> inscripciones = List.of(new InscripcionEvento(usuarioId, evento1),
+                new InscripcionEvento(usuarioId, evento2));
 
         when(inscripcionesRepository.getInscripcionesPorParticipante(usuarioId)).thenReturn(inscripciones);
